MWPW-124572: FG Copy action cleanup (#580)
